Fixed discount set as int? and not double? in interface and implementation

+++ server/BetterCalm/DataAccess/Repositories/AppointmentDiscountRepository.cs
@@ -19,7 +19,7 @@ namespace DataAccess.Repositories
             this.discounts = context.Set<AppointmentDiscount>();
         }
 
-        public AppointmentDiscount Get(int? discount)
+        public AppointmentDiscount Get(double? discount)
         {
             throw new NotImplementedException();
         }
